# Add _addmul1 ARM Neon implementation