--- a/zfec/fec.c
+++ b/zfec/fec.c
@@ -6,6 +6,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <assert.h>
 
@@ -64,6 +65,9 @@ static gf gf_mul_table[256][256];
 #define GF_MULC0(c) __gf_mulc_ = gf_mul_table[c]
 #define GF_ADDMULC(dst, x) dst ^= __gf_mulc_[x]
 
+/* gf_mul_table_16[c][x] = gf_mul_table[c][x << 4] */
+static gf gf_mul_table_16[256][16];
+
 /*
  * Generate GF(2**m) from the irreducible polynomial p(X) in p[0]..p[m]
  * Lookup tables:
@@ -83,6 +87,10 @@ _init_mul_table(void) {
 
   for (j = 0; j < 256; j++)
       gf_mul_table[0][j] = gf_mul_table[j][0] = 0;
+
+  for (i = 0; i < 256; i++)
+      for (j = 0; j < 16; j++)
+          gf_mul_table_16[i][j] = gf_mul_table[i][j << 4];
 }
 
 #define NEW_GF_MATRIX(rows, cols) \
@@ -174,6 +182,56 @@ _addmul1(register gf*restrict dst, const register gf*restrict src, gf c, size_t 
 
     GF_MULC0 (c);
 
+#if defined(ZFEC_USE_ARM_NEON) && UNROLL == 16
+    // Idea from https://botan.randombit.net zfec_vperm.cpp
+    
+    const gf * const gf_mulc_16 = gf_mul_table_16[c];
+    
+    /* align dst to 16 bytes */
+    while(sz && (uintptr_t)dst % 16) {
+        GF_ADDMULC (*dst, *src);
+        dst++;
+        src++;
+        sz--;
+    }
+    /* 
+     * for (; dst < lim; dst += 16, src += 16) {
+     *   // 16x parallel
+     *   for (int i = 0; i < 16; i++) {
+     *     dst[i] ^= __gf_mulc_[src[i] & 0x0f] ^ gf_mulc_16[src[i] >> 4];
+     *   }
+     * }
+     * dst should be aligned to 16 bytes
+     */
+    __asm__(
+        "vmov.i8  q0, #0x0f              \n\t"  // q0 = 0x0f0f...0f
+        "vld1.8   {q3}, [%[gf_mulc]]     \n\t"  // q3 = gf_mulc[0..15]
+        "vld1.8   {q8}, [%[gf_mulc_16]]  \n\t"  // q8 = gf_mulc_16[0..15]
+    "1:                                  \n\t"  //
+        "cmp      %[dst], %[lim]         \n\t"  //
+        "bhs      2f                     \n\t"  // while (dst < lim) {
+        "vld1.8   {q2}, [%[src]]!        \n\t"  //   q2 = src[0..15];  src += 16
+        "vld1.8   {q1}, [%[dst]:128]     \n\t"  //   q1 = dst[0..15]
+        "vand.8   q9, q2, q0             \n\t"  //   q9 = src & 0x0f0f...0f
+        "vtbl.8   d18, {q3}, d18         \n\t"  //
+        "vshr.u8  q10, q2, #4            \n\t"  //   q10 = src >> 4
+        "vtbl.8   d19, {q3}, d19         \n\t"  //   q9 = gf_mulc[q9]
+        "vtbl.8   d20, {q8}, d20         \n\t"  //
+        "vtbl.8   d21, {q8}, d21         \n\t"  //   q10 = gf_mulc_16[q10]
+        "veorq.8  q1, q9                 \n\t"  //
+        "veorq.8  q1, q10                \n\t"  //
+        "vst1.8   {q1}, [%[dst]:128]!    \n\t"  //   dst[0..15] ^= q9 ^ q10;  dst += 16
+        "b        1b                     \n\t"  //
+    "2:                                  \n\t"  // }
+        : [dst]"+r"(dst),
+          [src]"+r"(src)
+        : [lim]       "r"(lim),
+          [gf_mulc]   "r"(__gf_mulc_),
+          [gf_mulc_16]"r"(gf_mulc_16)
+        : "q0", "q1", "q2", "q3", "q8", "q9", "q10", "memory", "cc"
+    );
+
+#else
 #if (UNROLL > 1)                /* unrolling by 8/16 is quite effective on the pentium */
     for (; dst < lim; dst += UNROLL, src += UNROLL) {
         GF_ADDMULC (dst[0], src[0]);
@@ -198,6 +256,8 @@ _addmul1(register gf*restrict dst, const register gf*restrict src, gf c, size_t 
 #endif
     }
 #endif
+#endif
+
     lim += UNROLL - 1;
     for (; dst < lim; dst++, src++)       /* final components */
         GF_ADDMULC (*dst, *src);