@WojciechMigda 2022-08-04
I would suggest that with a hardware-specific optimizations existing github worflows should be extended with regression checks for ARM-compiled code. I am using `qemu` for that, but whatever works for you.