@gsr933 2022-08-06
> I would suggest that with hardware-specific optimizations existing github worflows should be extended with regression checks for ARM-compiled code. I am using `qemu` for that, but whatever works for you.

Added in 52fe7fa