@sajith 2022-08-09
Thank you for the PR!  I clicked the button that says "Approve and run" the workflow.

I seem to be a zfec maintainer, and I can't speak for other maintainers, but: I do not know how to read ARM assembly, and I am not going to have the time to learn it any time soon.  I don't even know what ARM Neon is, and how it is different from regular ARM!  But it might be useful if you could explain a few things:

* The primary consumer of zfec is Tahoe-LAFS project, which maintains zfec.  As you can see there's just enough activity in zfec to keep it going to suit Tahoe-LAFS project's needs, because we do have bandwidth constraints.  How could this PR help Tahoe-LAFS project?  I do not think we have ever heard from anyone using Tahoe-LAFS on ARM Neon.

* Why are you adding this to zfec?  Who will it help?  Zfec hasn't had any inline assembly up to now, and I am apprehensive about adding it now.

* What about ongoing maintenance?  Who will help us fix issues with ARM Neon bits if/when they are uncovered in the future?

* We have been talking about moving Tahoe-LAFS org to GitLab.  So another concern I have is that we will have to do extra work if/when we set up CI there.

Is it important to you that this is merged?