@ribbles 2022-08-10
@sajith the code sits behind a feature-flag so it wont have any impact on existing users of the module, unless they want hardware acceleration. It's very common for FEC libraries to use hardware acceleration as they are well suited to benefit from dedicated CPU instructions like SIMD and ADDMUL, much like AES hardware acceleration for cryptography.

- ARM is the CPU in a Raspberry PI and the new Apple M1 MacBook
- The purpose of the PR is to reduce the time taken to process a block of data
- The GutHub CI portion of the commit was done entirely at the request of user WojciechMigda.
- It's quite possible in the future that someone will undertake enabling hardware acceleration for Intel/AMD processors - that's the power of open source git repos!