@sajith 2022-08-10
@ribbles Since users of Tahoe-LAFS haven't asked that zfec be sped up, speeding up zfec isn't a priority right now.  If we absolutely must speed zfec up, we should probably consider something like OpenMP first, and let compilers handle optimizations.  But even that ever has not been a priority.

The existing pieces in C itself is hard to understand, and adding assembly would complicate it further.  I personally wouldn't be the one to click "the merge PR" button, because I don't know how to review this code and it would be irresponsible of me to merge code that I do not understand.  It would be especially irresponsible because I don't spend much time on Tahoe-LAFS these days.  There is a maintenance cost to adding any new code, especially assembly, and I don't want to burden other maintainers.  Unless they want to burden themselves, of course.  Which they seem reluctant to do.

I brought this PR up in Tahoe-LAFS IRC channel.  The consensus is that this PR is going to be hard to merge, without spending considerable time and effort in reviewing the code.

Zfec's git history stars in 2007, and it used other version control systems before.  No one has added hand-written AMD/Intel SIMD optimizations all those years. If anyone is going to do that, I hope they have a persuasive case, and that they try to persuade before they spend any effort with a PR.

Another power of free and open source repositories is that you can maintain a fork yourself. :-)