@dan-glass 2022-08-10
When `zfec` was added to Pypi it became something much bigger than just the Tahoe-LAFS project. `zfec` pip downloads last month were 2,863 vs. 1,481 for `Tahoe-LAFS` - however maybe that's not representative.