@sajith 2022-08-10
Even if zfec turns out be far more popular and important than Tahoe-LAFS, and even if zfec maintainers (hypothetically) could be convinced that adding assembly to zfec is a good idea, it remains that somebody has to review and merge this PR.  That somebody is not me.  I am not qualified to review this PR.